feat: rewrite common js module to it's equivalent in ECMAScript notation
