+++ src/modules/esm.mjs
@@ -1,0 +1,42 @@
+import {sep, dirname} from "node:path";
+import {release, version} from "node:os"
+import {createServer as createHttpServer} from "node:http";
+import {fileURLToPath} from "node:url";
+import {createRequire} from "module";
+import './files/c.js';
+
+const require = createRequire(import.meta.url)
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = require('./files/a.json');
+} else {
+    unknownObject = require('./files/a.json');
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+console.log(`Path to current file is ${fileURLToPath(import.meta.url)}`);
+console.log(`Path to current directory is ${dirname(fileURLToPath(import.meta.url))}`);
+
+const myServer = createHttpServer((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+    unknownObject,
+    myServer,
+};
